% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{do_gbm}
\alias{do_gbm}
\title{Run Gradient Boosting Model}
\usage{
do_gbm(data, features, outcome, tg)
}
\arguments{
\item{data}{data.frame the dataset including features and outcomes}

\item{features}{character the column names of features (X) in data}

\item{outcome}{character the columns names of outcome (y) in data}

\item{tg}{tuning grid for caret}
}
\value{
the list of feature score (gbm_scores) and models (gbm_model)
}
\description{
This function train and return the gradient boosting model using the package "caret"
}
