% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_function_v2.R
\name{do_gbm_v2}
\alias{do_gbm_v2}
\title{Run the gradient boosting model}
\usage{
do_gbm_v2(data, features, outcome, tg)
}
\arguments{
\item{data}{data.frame the dataset including features and outcomes}

\item{features}{character the column names of features (X) in data}

\item{outcome}{character the columns names of outcome (y) in data}

\item{tg}{tuning grid for caret}
}
\value{
the list of model score (metric), feature score (scores) and models (model)
}
\description{
This function train and return the gradient boosting model using "caret" and "gbm"
}
