% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{do_rf}
\alias{do_rf}
\title{Run RandomForest}
\usage{
do_rf(data, features, outcome)
}
\arguments{
\item{data}{data.frame the dataset including features and outcomes}

\item{features}{character the column names of features (X) in data}

\item{outcome}{character the columns names of outcome (y) in data}
}
\value{
the list of feature score (rf_scores) and models (rf_model)
}
\description{
This function train and return the random forest using the package "randomForest"
}
