% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_function_v2.R
\name{get_scores}
\alias{get_scores}
\title{Making DataFrame for scores}
\usage{
get_scores(ml_result_score_list, method)
}
\arguments{
\item{ml_result_score_list}{list list of importance scores of models}

\item{method}{vector vector of character with methods name}
}
\value{
the list of model score (metric), feature score (scores) and models (model)
}
\description{
This function will return the tran and test functions
}
