% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_Regression.R
\name{L1_penalty}
\alias{L1_penalty}
\title{Calculating L1-penalty}
\usage{
L1_penalty(beta_temp)
}
\arguments{
\item{beta_temp}{a vector of parameters}
}
\value{
L1 penalty
}
\description{
This function evaluate the L1 penalty for current beta
}
