% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_Regression.R
\name{lasso_bootstrap_inference}
\alias{lasso_bootstrap_inference}
\title{Finding optimal Lasso parameter and Lasso coefficient}
\usage{
lasso_bootstrap_inference(X, y, B = 100)
}
\arguments{
\item{X}{a matrix of predictors}

\item{y}{a vector of response}

\item{B}{a number of bootstrap samples}
}
\value{
list of results of Lasso Bootstap Inference including Lasso Coefficients and Feature Scores.
}
\description{
This function estimate optimal Lasso parameter and Lasso coefficient using Cross Validation
}
