% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_Regression.R
\name{lasso_optimal}
\alias{lasso_optimal}
\title{Finding optimal Lasso parameter and Lasso coefficient}
\usage{
lasso_optimal(X, y, max_iter = 10000, eps = 10^(-10), cv_fold = 5)
}
\arguments{
\item{X}{a matrix of predictors}

\item{y}{a vector of response}

\item{max_iter}{Maximum number of iteration}

\item{eps}{epsilon used at convergence criterion}

\item{cv_fold}{number of folds}
}
\value{
list of minimal and 1se Lasso parameter (min_lambda,se1_lambda) and Lasso coefficient (min_beta,se1_beta)
}
\description{
This function estimate optimal Lasso parameter and Lasso coefficient using Cross Validation
}
