% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_Regression.R
\name{Lasso_regression}
\alias{Lasso_regression}
\title{Estimating Lasso Coefficient}
\usage{
Lasso_regression(X, y, lambda = 1, max_iter = 10000, eps = 10^(-8))
}
\arguments{
\item{X}{a matrix of predictors}

\item{y}{a vector of response}

\item{lambda}{lasso penalty parameter}

\item{max_iter}{Maximum number of iteration}

\item{eps}{epsilon used at convergence criterion}
}
\value{
list of optimal (Beta) and beta candidate per iteration (Optim_Trajectory)
}
\description{
This function estimate optimal Lasso Coefficient using ISTA and coordinate descent
}
