% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_Regression.R
\name{RegressionNegLogLikelihood}
\alias{RegressionNegLogLikelihood}
\title{Calculating negative log likelihood}
\usage{
RegressionNegLogLikelihood(beta, X, y)
}
\arguments{
\item{beta}{a vector of parameters}

\item{X}{a matrix of predictors}

\item{y}{a vector of response}
}
\value{
negative log likelihood
}
\description{
This function evaluate the negative loglikelihood of given beta, X and y for ordinary linear models
}
