% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_Regression.R
\name{soft_thresholding}
\alias{soft_thresholding}
\title{SoftThresholding Function}
\usage{
soft_thresholding(x, lambda)
}
\arguments{
\item{x}{input}

\item{lambda}{Lasso Parameter}
}
\value{
output of soft thresholding function
}
\description{
This function is used at optimazing Lasso Regression
}
